requirements.txt
flask==2.0.1
opencv-python-headless==4.5.3.56
spotipy==2.19.0
python-dotenv==0.19.0
numpy==1.21.2
gunicorn==20.1.0
Pillow==9.0.0
mtcnn==0.1.1
transformers==4.30.2
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
nltk==3.8.1
